+++ rust/stackablectl/src/cmds/operator.rs
@@ -73,7 +73,7 @@ pub struct OperatorListArgs {
 
     #[arg(
         long,
-        long_help = "Source the charts from either a OCI registry or from Nexus repositories.",
+        long_help = "Source the charts from either a OCI registry or from index.yaml-based repositories.",
         value_enum, default_value_t = Default::default()
     )]
     chart_source: ChartSourceTypeArg,
@@ -90,7 +90,7 @@ pub struct OperatorDescribeArgs {
 
     #[arg(
         long,
-        long_help = "Source the charts from either a OCI registry or from Nexus repositories.", value_enum, default_value_t = Default::default()
+        long_help = "Source the charts from either a OCI registry or from index.yaml-based repositories.", value_enum, default_value_t = Default::default()
     )]
     chart_source: ChartSourceTypeArg,
 }
@@ -123,7 +123,7 @@ Use \"stackablectl operator describe <OPERATOR>\" to get available versions for 
 
     #[arg(
         long,
-        long_help = "Source the charts from either a OCI registry or from Nexus repositories.",
+        long_help = "Source the charts from either a OCI registry or from index.yaml-based repositories.",
         value_enum, default_value_t = Default::default()
     )]
     chart_source: ChartSourceTypeArg,