+++ rust/stackablectl/src/cmds/release.rs
@@ -85,7 +85,7 @@ pub struct ReleaseInstallArgs {
 
     #[arg(
         long,
-        long_help = "Source the charts from either a OCI registry or from Nexus repositories.",
+        long_help = "Source the charts from either a OCI registry or from index.yaml-based repositories.",
         value_enum, default_value_t = Default::default()
     )]
     chart_source: ChartSourceTypeArg,